Brings 34 commits including Stage 3-5 close/advance, worldgen fixes,
Stage 2 lifecycle parity, and docs improvements.

+++ src/main/java/thaumcraft/codechicken/lib/math/MathHelper.java
@@ -88,7 +88,7 @@ public class MathHelper {
         return a == b ? 0 : (a < b ? -1 : 1);
     }
 
-    public static double compare(double a, double b) {
+    public static int compare(double a, double b) {
         return a == b ? 0 : (a < b ? -1 : 1);
     }
